feat: persist metric filter preference
